feat: use date values over index for infinite scroll
Instead of calculating the date from the index, as was currently done,
we now calculate the dates whenever the `start`/`end` store limits
change. This approach simplifies the calculations as we now only rely on
dates and such dates are not coupled to the store's dateRange
implementation.
Ideally, once the dnd is finished, we should look into virtualizing the
lists since the user can scroll infinitely...